Conversion/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==4.0.1; python_version >= '3.6'
certifi==2022.9.24; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
flask-bcrypt==1.0.1
flask==2.2.2
idna==3.4; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
pymysql==1.0.2
requests==2.28.1
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
werkzeug==2.2.2; python_version >= '3.7'


